Move log to core
