+++ packages/core/values/defaults.js
@@ -1,4 +1,4 @@
-import { advise } from '../../common/log'
+import { advise } from '../log'
 import settings from './settings'
 
 const onReadyDeprecated = (messageData) => {
